feat: Phase 2 DI, settings, secrets store, and package layout
Introduce create_app() with app.state DI (routes no longer use set_*
globals), pydantic-settings SpectraXSettings, SecretsStore
(file/keyring/memory) behind the credentials facade, and package moves
into recording/, detection/, and api/. Add serve/doctor CLI entry points,
deploy/systemd units, and tests for config/secrets/app factory. Phase 0
characterization and auth suites remain the regression net.
