+++ pyproject.toml
@@ -24,6 +24,7 @@ dependencies = [
     "argon2-cffi==23.1.0",
     "python-multipart==0.0.20",
     "pydantic==2.11.4",
+    "pydantic-settings==2.9.1",
     "Pillow==11.2.1",
     "requests==2.32.3",
     "tqdm==4.67.1",
