+++ src/spectrax/api/__init__.py
@@ -1,0 +1,1 @@
+"""HTTP API layer (deps + routers)."""
