+++ src/spectrax/detection/__init__.py
@@ -1,0 +1,6 @@
+"""Detection subsystem."""
+
+from .detector import DetectorManager, RTSPObjectDetector
+from .config import DetectorConfig
+
+__all__ = ["DetectorManager", "RTSPObjectDetector", "DetectorConfig"]
