+++ src/spectrax/detection/backends/__init__.py
@@ -1,0 +1,1 @@
+"""Detection backends extension point (empty seam; ONNX/NCNN deferred)."""
