+++ src/spectrax/mediamtx/__init__.py
@@ -1,0 +1,1 @@
+"""MediaMTX config generation and process management."""
