+++ src/spectrax/recording/__init__.py
@@ -1,0 +1,5 @@
+"""Recording subsystem (DB + manager)."""
+
+from .db import RecordingsAPI
+
+__all__ = ["RecordingsAPI"]
